start replacing nbody with almanac
