+++ learning/data/iss-example.toml
@@ -1,0 +1,51 @@
+sequence = ["iss", "iss_cond"]
+
+# This example is taken from Poliastro
+
+[state.iss_init]
+x = 8.59072560e2
+y = -4.13720368e3
+z = 5.29556871e3
+vx = 7.37289205
+vy = 2.08223573
+vz = 4.39999794e-1
+frame = "EME2000"
+epoch = "2013-03-18T12:00:00.0 UTC"
+unit_position = "km"  # Default value if unspecified
+unit_velocity = "km/s"  # Default value if unspecified
+
+[orbital_dynamics.iss]
+integration_frame = "EME2000"
+initial_state = "iss_init"
+point_masses = ["Earth"]
+accel_models = ["my_models"]
+
+[spacecraft.iss]
+dry_mass = 100.0
+fuel_mass = 20.0
+orbital_dynamics = "iss"
+
+[propagator.iss]
+dynamics = "iss"
+stop_cond = "30.0 days"
+output = "my_csv"
+
+[accel_models.my_models.harmonics.jgm3_70x70]
+frame = "EME2000"
+degree = 70
+order = 70
+file = "data/JGM3.cof.gz"
+
+[output.my_csv]
+filename = "./data/iss.csv"
+headers = ["epoch:GregorianUtc", "x", "y", "z", "vx", "vy", "vz", "rmag:Luna"]
+
+[propagator.iss_cond]
+dynamics = "iss"
+stop_cond = "apo"
+output = "my_csv"
+tolerance = 1e-9
+
+[conditions.apo]
+event = "apoapsis"
+search_until = "2013-03-18T14:00:00.0 UTC"
